Move check_user decorator to views.py.
